Refactor code - separate module functionality/scope, write dry code
Previously, the modules looked a little bit like spaghetti code, and
elements that didn't need to be exported/imported were being structured
to do so. Now there's a much cleaner entry point to index.js.

In addition, the logical structure of the 4 JS modules modules have been
changed. Whereas previously the primary entry-point 'index.js' acted as
simply a juncture for the other 3 modules to import their respective
functions, the 'index.js' file now acts more like the C&C center of
operations, calling each of the other 3 modules as necessary, with
the core logic being written in this file.